src/backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
supabase==2.28.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Async job queue
celery==5.4.0
redis==5.1.0

# Video processing - may not be needed
yt-dlp==2024.11.4

# OCR (text from video frames)
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
Pillow==10.4.0

# Google AI
google-generativeai
google-cloud-speech
google-cloud-texttospeech

# Utilities
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.12